Reworking program structure and data structure
Program will be run with command line arguments for the following
operations:
data collection,
training,
playing

There is a need to rework the models input data due to some
inconsistencies realized during testing. The model currently whatever at
that given moment a user had pressed down paired with a single screen
capture. This is saying that our model should press buttons we used
before we saw the frame when seeing a given frame. The mouse movement
also only keeps track of the last dx and dy which occurs every
mouse_move event. We should rather train the model to look for what keys
were pressed after a given frame and the frame that occured after that.
We should also try to give the model the concept of state by using an
array mapping how long a key has been pressed down to into a range (0,1)
which should solve the problem with hooks failing to let go and the
model not staying on track.